requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-socketio>=5.11.0
